+++ RxCocoa/Traits/Driver/Driver+Subscription.swift
@@ -7,6 +7,7 @@
 //
 
 import RxSwift
+import RxRelay
 
 private let errorMessage = "`drive*` family of methods can be only called from `MainThread`.\n" +
 "This is required to ensure that the last replayed `Driver` element is delivered on `MainThread`.\n"
