+++ RxRelay.podspec
@@ -1,11 +1,11 @@
 Pod::Spec.new do |s|
-  s.name             = "RxCocoa"
+  s.name             = "RxRelay"
   s.version          = "5.0.0"
-  s.summary          = "RxSwift Cocoa extensions"
+  s.summary          = "Relays for RxSwift - PublishRelay and BehaviorRelay"
   s.description      = <<-DESC
-* UI extensions
-* NSURL extensions
-* KVO extensions
+* PublishRelay
+* BehaviorRelay
+* Binding overloads
                         DESC
   s.homepage         = "https://github.com/ReactiveX/RxSwift"
   s.license          = 'MIT'
@@ -19,11 +19,8 @@ Pod::Spec.new do |s|
   s.watchos.deployment_target = '3.0'
   s.tvos.deployment_target = '9.0'
 
-  s.source_files          = 'RxCocoa/**/*.{swift,h,m}', 'Platform/**/*.swift'
-  s.exclude_files         = 'RxCocoa/Platform/**/*.swift', 'Platform/AtomicInt.swift'
+  s.source_files          = 'RxRelay/**/*.{swift,h,m}'
 
   s.dependency 'RxSwift', '~> 5'
-  s.dependency 'RxRelay', '~> 5'
-
   s.swift_version = '5.0'
 end
