+++ RxRelay/BehaviorRelay.swift
@@ -1,6 +1,6 @@
 //
 //  BehaviorRelay.swift
-//  RxCocoa
+//  RxRelay
 //
 //  Created by Krunoslav Zaher on 10/7/17.
 //  Copyright © 2017 Krunoslav Zaher. All rights reserved.
