+++ RxRelay/Observable+Bind.swift
@@ -1,0 +1,81 @@
+//
+//  Observable+Bind.swift
+//  RxRelay
+//
+//  Created by Shai Mishali on 09/04/2019.
+//  Copyright © 2019 Krunoslav Zaher. All rights reserved.
+//
+
+import RxSwift
+
+extension ObservableType {
+    /**
+     Creates new subscription and sends elements to publish relay.
+
+     In case error occurs in debug mode, `fatalError` will be raised.
+     In case error occurs in release mode, `error` will be logged.
+
+     - parameter to: Target publish relay for sequence elements.
+     - returns: Disposable object that can be used to unsubscribe the observer.
+     */
+    public func bind(to relay: PublishRelay<E>) -> Disposable {
+        return self.subscribe { e in
+            switch e {
+            case let .next(element):
+                relay.accept(element)
+            case let .error(error):
+                rxFatalErrorInDebug("Binding error to publish relay: \(error)")
+            case .completed:
+                break
+            }
+        }
+    }
+
+    /**
+     Creates new subscription and sends elements to publish relay.
+
+     In case error occurs in debug mode, `fatalError` will be raised.
+     In case error occurs in release mode, `error` will be logged.
+
+     - parameter to: Target publish relay for sequence elements.
+     - returns: Disposable object that can be used to unsubscribe the observer.
+     */
+    public func bind(to relay: PublishRelay<E?>) -> Disposable {
+        return self.map { $0 as E? }.bind(to: relay)
+    }
+
+    /**
+     Creates new subscription and sends elements to behavior relay.
+
+     In case error occurs in debug mode, `fatalError` will be raised.
+     In case error occurs in release mode, `error` will be logged.
+
+     - parameter to: Target behavior relay for sequence elements.
+     - returns: Disposable object that can be used to unsubscribe the observer.
+     */
+    public func bind(to relay: BehaviorRelay<E>) -> Disposable {
+        return self.subscribe { e in
+            switch e {
+            case let .next(element):
+                relay.accept(element)
+            case let .error(error):
+                rxFatalErrorInDebug("Binding error to behavior relay: \(error)")
+            case .completed:
+                break
+            }
+        }
+    }
+
+    /**
+     Creates new subscription and sends elements to behavior relay.
+
+     In case error occurs in debug mode, `fatalError` will be raised.
+     In case error occurs in release mode, `error` will be logged.
+
+     - parameter to: Target behavior relay for sequence elements.
+     - returns: Disposable object that can be used to unsubscribe the observer.
+     */
+    public func bind(to relay: BehaviorRelay<E?>) -> Disposable {
+        return self.map { $0 as E? }.bind(to: relay)
+    }
+}
