--- Sources/RxCocoa/BehaviorRelay.swift
@@ -1,1 +1,0 @@
-../../RxCocoa/Traits/BehaviorRelay.swift
