--- Sources/RxCocoa/PublishRelay.swift
@@ -1,1 +1,0 @@
-../../RxCocoa/Traits/PublishRelay.swift
